fix type error

+++ src/components/Withdraw.js
@@ -10,17 +10,14 @@ class Withdraw extends React.Component {
         this.state = {
             _id: this.props.accounts._id,
             amount: ''
-            //console.log('hello')
-            //this.state = { _id: 1 };
-            // console.log(this.state)
         }
     }
 
     onFormSubmit = (event) => {
         //console.log(this.state)
         event.preventDefault();
-        this.props.withdraw(this.state._id, parseInt(this.state.amount));
-        this.props.addTransactions('withdraw', parseInt(this.state.amount), this.state.name)
+        this.props.withdraw(this.state._id, parseFloat(this.state.amount));
+        this.props.addTransactions('withdraw', parseFloat(this.state.amount), this.state.name)
         console.log(this.state)
         //console.log(parseInt(this.state.amount));
         this.setState({amount:''});
